fixes typo on event handler (#290)
The function for the `myEventType` specifies two arguments (state and event). There appears to be a typo with the variable `evnt` (should be) `event`. The functions appears to increment the `count` by one on the incoming `state` variable however the code uses the variable `s` which will causes an issue as the `s` variable is undefined.